Add custom Buffer module to reduce bundle size
Browserify and Webpack shims for node Buffer include a lot of methods that will be likely never used by this lib.
This commit add a stripped version of the original Buffer shim, with only the methods strictly needed.

Bundle size is reduced by ~50%, resulting in a footprint of only ~7Kb (minified and gzipped).